backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-genai==0.3.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.3
